apps/backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
pydantic>=2.9,<3.0
pydantic-settings>=2.5,<3.0
google-cloud-firestore>=2.18,<3.0
firebase-admin>=6.6,<7.0
google-genai>=1.0,<2.0
google-adk>=0.1,<1.0
Pillow>=10.0,<12.0
python-dotenv>=1.0,<2.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
